Fixed will-filter problem by changing all decimal datatypes in migrations to float datatypes.

+++ db/migrate/20120801152757_add_discount_price_and_spot_check_rating_to_job.rb
@@ -1,6 +1,6 @@
 class AddDiscountPriceAndSpotCheckRatingToJob < ActiveRecord::Migration
   def change
-    add_column :jobs, :discount_price, :decimal
+    add_column :jobs, :discount_price, :float
     add_column :jobs, :rating_superviser, :integer
     rename_column :jobs, :rating, :rating_client
   end
